% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addValuesToDB}
\alias{addValuesToDB}
\title{Add User's input to Postgres database}
\usage{
addValuesToDB(tableName, newValuesList, dbConnection)
}
\arguments{
\item{tableName}{A \code{character} with the name of table to which the values will be inputted}

\item{newValuesList}{A \code{list} of values to add; the length of the list must be the same as number of columns
of the table minus 1 (except 'id' column)}

\item{dbConnection}{A connection to the Postgres database}
}
\description{
Add User's input to database table imputing the missing id column with the next possible values
}
\details{
This function add the input provided by the User in the App. The function assumes that primary key
of the each table in database is named 'id' and cannot be added by the User. The 'id' column is
being added inside the function after the next possible value of 'id' is taken from the database.
}
\author{
Alina Tselinina \href{mailto:tselinina@gmail.com}{tselinina@gmail.com}
}
