% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{bands}
\alias{bands}
\title{Example bands data}
\format{
A \code{data.frame} of 2 columns and 10 rows
\describe{
\item{id}{id value}
\item{name}{character value with the name of a band}
}
}
\source{
Simulated data
}
\usage{
bands
}
\description{
A dataset contains example of 'bands' table
}
\keyword{datasets}
