% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{events}
\alias{events}
\title{Example music events data}
\format{
A \code{data.frame} of 4 columns and 10 rows
\describe{
\item{id}{id value}
\item{name}{character value with the name of an event}
\item{musician_id}{id value of a musician took part in the event as a member of band_id band}
\item{band_id}{id value of a band took part in the event}
}
}
\source{
Simulated data
}
\usage{
events
}
\description{
A dataset contains example of 'events' table
}
\keyword{datasets}
