% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getConnectionToDB}
\alias{getConnectionToDB}
\title{Getting connection to the Postgres DB}
\usage{
getConnectionToDB()
}
\value{
Connection to database
}
\description{
Getting connection to the Postgres DB from environmental settings
}
\details{
This function gets connections to the DB based on environmental variables
which contains information about the name of database (DB_NAME), user (DB_USER),
password (DB_PASSWORD), port (DB_PORT), host (DB_HOST) and schema (DB_SCHEMA).
}
\author{
"Alina Tselinina \href{mailto:tselinina@gmail.com}{tselinina@gmail.com}"
}
