% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getInfoAboutMusician}
\alias{getInfoAboutMusician}
\title{Send queries to Postgres database to get relations}
\usage{
getInfoAboutMusician(name, dbConnection)
}
\arguments{
\item{name}{A \code{character} which is the full name of musician}

\item{dbConnection}{A connection to Postgres database}
}
\value{
A \code{list} of two \code{data.frame} with the relations; the first
\code{data.frame} consists of one column 'band' and describes the musician-bands
relation; the second \code{data.frame} consists of two 'name' and 'surname'
columns and describe musician-musicians relation;
}
\description{
Send queries to Postgres database to get relations between bands and musicians
}
\details{
This function sends the queries to database to get musician-bands and
musician-musicians relations
}
\author{
"Alina Tselinina \href{mailto:tselinina@gmail.com}{tselinina@gmail.com}"
}
