% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inputIsCorrect}
\alias{inputIsCorrect}
\title{Check User's input}
\usage{
inputIsCorrect(listOfInput, data, message = FALSE)
}
\arguments{
\item{listOfInput}{A \code{list} containing inputted values to check}

\item{data}{A reactive \code{data.frame} with all current values}

\item{message}{\code{TRUE/FALSE} value to choose the function output (\code{TRUE} for getting message
with instruction to correct the input; \code{FALSE} for getting \code{TRUE/FALSE} status of
input check); \code{FALSE} by default}
}
\value{
Returns \code{TRUE} if all input is correct, otherwise, returns \code{FALSE}.
If \code{messages=TRUE} the function returns the message to the User with appropriate instruction.
}
\description{
Check the correctness of the input provided by the User
}
\details{
This function checks if the numeric and character input is not empty
as well as checks numeric values for being integers.
}
\examples{
inputIsCorrect(list("    ", 1.5, 2))
inputIsCorrect(list("    ", 1.5, 2), TRUE)
inputIsCorrect(list("Hangout Music Festival", 1L, 3L))
}
\author{
Alina Tselinina \href{mailto:tselinina@gmail.com}{tselinina@gmail.com}
}
