% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{launch_app}
\alias{launch_app}
\alias{\link{musicBox_ui}}
\alias{\link{musicBox_server}}
\title{Launch MusicBox Shiny App}
\usage{
launch_app()
}
\description{
Launch MusicBox Shiny App
}
\details{
This function launches MusicBox Shiny App based on the data from Postgres database
which must be connected to R before launching.
}
\author{
Alina Tselinina \href{mailto:tselinina@gmail.com}{tselinina@gmail.com}
}
