% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicBox.R
\docType{package}
\name{musicBox}
\alias{musicBox}
\alias{musicBox-package}
\title{musicBox: A package with a Shiny App inside}
\description{
This package provides functionality for running a Shiny App which allows for adding
new values to Postgres database containing data with musicians, bands and events
interactively and for showing relations between musicians and bands
based on User choice of interest.
}
\author{
Alina Tselinina \email{tselinina@gmail.com}
}
