% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{musicBox_ui}
\alias{musicBox_ui}
\title{UI function for MusicBox Shiny App}
\usage{
musicBox_ui()
}
\description{
UI function for MusicBox Shiny App
}
\details{
UI function generates the front-end for MusicBox Shiny App. The function sources
two small UI R scripts, one per each tab of the App
}
\author{
"Alina Tselinina \href{mailto:tselinina@gmail.com}{tselinina@gmail.com}"
}
