% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicians.R
\docType{data}
\name{musicians}
\alias{musicians}
\title{musicians data set}
\format{
A \code{data.frame} with 3 columns
\describe{
\item{id}{id value}
\item{name}{character value with the first name of a musician}
\item{surname}{character value with the second name of a musician}
}
}
\source{
Simulated Data
}
\usage{
musicians
}
\description{
The data represents an example table with musicians
}
\keyword{datasets}
