% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepareStoryAboutBand}
\alias{prepareStoryAboutBand}
\title{Generate human-friendly story for Band}
\usage{
prepareStoryAboutBand(name, tableList)
}
\arguments{
\item{name}{A \code{character} which is the name of band}

\item{tableList}{A \code{list} with two data frames; the first \code{data.frame} consists
of one \code{character} column called 'band'; the second \code{data.frame} consists of
two character columns: 'name' and 'surname'}
}
\value{
A \code{list} of two \code{character} with the stories; the first story
describes the band-bands relations; the second - band-musicians relations
}
\description{
Generate human-friendly story for Band based on the tables
}
\details{
This function generates a story based on the list of two data frames. One
data frame contains the band names which the band played with. The second
table contains the name and surname of musicians who played in the band.
The story is being generated depending on the number of rows.
}
\author{
Alina Tselinina \href{mailto:tselinina@gmail.com}{tselinina@gmail.com}
}
