% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepareStoryAboutMusician}
\alias{prepareStoryAboutMusician}
\title{Generate human-friendly story for Musician}
\usage{
prepareStoryAboutMusician(name, tableList)
}
\arguments{
\item{name}{A \code{character} which is the full name of musician}

\item{tableList}{A \code{list} with two data frames; the first \code{data.frame} consists
of one \code{character} column called 'band'; the second \code{data.frame} consists of
two character columns: 'name' and 'surname'}
}
\value{
A \code{list} of two \code{character} with the stories; the first story
describes the musician-bands relations; the second - musician-musicians relations
}
\description{
Generate human-friendly story for Musician based on the tables
}
\details{
This function generates a story based on the list of two data frames. One
data frame contains the band names which the musician played in. The second
table contains the name and surname of musicians who played with the musician.
The story is being generated depending on the number of rows.
}
\examples{
prepareStoryAboutMusician('John Lenon',
list(data.frame(list('band'='The Rolling Stone')),
data.frame(list('name'='Paul', 'surname'='McCartney'))))

}
\author{
Alina Tselinina \href{mailto:tselinina@gmail.com}{tselinina@gmail.com}
}
