% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_show_relations.R
\name{showInfo}
\alias{showInfo}
\title{Generating front-end with the stories}
\usage{
showInfo(input, output, session, typeOfObject, valueOfObject, db_connection)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{environment that can be used to access information and functionality}

\item{typeOfObject}{A \code{character} with the User's choice (e.g, 'Musician', 'Band')}

\item{valueOfObject}{A \code{character} with the full name of musician or band name}

\item{db_connection}{A connection to Postgres database}
}
\description{
Generating front-end with the stories for band/musician chosen by the User
}
\details{
Server function of the module which generates stories for band or musician
depending on the User's choice. The text of stories depending on the number
of found band and musician relations.
}
