% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_show_relations.R
\name{showInfoUI}
\alias{showInfoUI}
\title{Generating front-end with the stories}
\usage{
showInfoUI(id)
}
\arguments{
\item{id}{A \code{character} with unique ID}
}
\description{
Generating front-end with the stories for band/musician chosen by the User
}
\details{
UI function of the module which generates stories for band or musician
depending on the User's choice. The text of stories depending on the number
of found band and musician relations.
}
