% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_show_input_for_table.R
\name{showInput}
\alias{showInput}
\title{Generating front-end with the dynamic Input}
\usage{
showInput(input, output, session, data)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{environment that can be used to access information and functionality}

\item{data}{reactive table}
}
\description{
Generating front-end with the dynamic Input for the table chosen by the User
}
\details{
Server function of module which generates dynamic Input based
on the chosen table.
}
