% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_show_input_for_table.R
\name{showInputUI}
\alias{showInputUI}
\title{Generating front-end with the dynamic Input}
\usage{
showInputUI(id)
}
\arguments{
\item{id}{A \code{character} with unique ID}
}
\description{
Generating front-end with the dynamic Input for the table chosen by the User
}
\details{
UI function of the module which generates dynamic Input based on the name and
type of the columns of the chosen table. Only character and integer columns are
taken
}
