% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\docType{data}
\name{bands}
\alias{bands}
\title{bands data set}
\format{
A \code{data.frame} with 2 columns
\describe{
\item{id}{id value}
\item{name}{character value with the name of a band}
}
}
\source{
Simulated Data
}
\usage{
bands
}
\description{
The data represents an example table with bands
}
\keyword{datasets}
