% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{events}
\alias{events}
\title{events data set}
\format{
A \code{data.frame} with 4 columns
\describe{
\item{id}{id value}
\item{name}{character value with the name of an event}
\item{musician_id}{id value of a musician took part in the event}
\item{band_id}{id value of a band took part in the event}
}
}
\source{
Simulated Data
}
\usage{
events
}
\description{
The data represents an example table with events
}
\keyword{datasets}
