% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{package}
\name{musicBox}
\alias{musicBox}
\title{musicBox: A package with a Shiny App inside}
\description{
this package provides functionality for running Shiny App which allows for
adding new values to Postgres database containing data with musicians, bands and
music events interactively and for showing relations between musicians and bands
based on User's choice of interest.
}
\author{
Alina Tselinina \email{tselinina@gmail.com}
}
