% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{musicians}
\alias{musicians}
\title{Example musicians data}
\format{
A \code{data.frame} of 3 columns and 10 rows
\describe{
\item{id}{id value}
\item{name}{character value with the first name of a musician}
\item{surname}{character value with the surname of musician}
}
}
\source{
Simulated data
}
\usage{
musicians
}
\description{
A dataset contains example of 'musicians' table
}
\keyword{datasets}
